app/mcmc_stats_server/pyproject.toml
[tool.poetry]
name = "chainsail-mcmc-stats-server"
version = "0.1.0"
description = "Web app that serves MCMC stats"
authors = ["Simeon Carstens <simeon.carstens@tweag.io>"]

[tool.poetry.dependencies]
python = "^3.8"
chainsail-common = { path = "../../lib/common", develop = true }
Flask = "^1.1.2"
PyYAML = "^5.3.1"
uWSGI = "^2.0.19"
click = "^8.0.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
black = "^20.8b1"
flake8 = "^3.8.4"
pylint = "^2.6.0"
mypy = "^0.790"
pytest-dotenv = "^0.5.2"
pytest-cov = "^2.10.1"
isort = "^5.7.0"

[tool.pytest.ini_options]
minversion = "6.0"
env_override_existing_values = 1
env_files = [".env"]

[tool.coverage.run]
omit = ["*test*"]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

